+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1 — May 24th, 2025
 
 - Fix how `english` library is required.
+- Fix linting on CI.
 
 ## 0.1.0 — May 24th, 2025